auth-service/requirement.txt
bcrypt==3.2.2
cffi==1.15.0
click==8.1.3
colorama==0.4.4
cryptography==37.0.2
Flask==2.1.2
importlib-metadata==4.11.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
passlib==1.7.4
psycopg2-binary==2.9.3
pycparser==2.21
PyJWT==2.4.0
python-dotenv==0.20.0
Werkzeug==2.1.2
zipp==3.8.0
